Updated the model loading Section
